who can make learn in group network App project
Group learn network App project, it is learn everywhere in the world, For University at home and at school connect together with class to 
talk about lesson, face to face, we will make New App = Learn In Group Network, for Cambodia and over the world. So, Who in group to make it to well.